XW-1304 | Separate Map and ImageMedia components
