+++ front/main/app/components/article-content.js
@@ -5,7 +5,7 @@ import GalleryMediaComponent from './gallery-media';
 import VideoMediaComponent from './video-media';
 import ImageMediaComponent from './image-media';
 import InfoboxImageCollectionComponent from './infobox-image-collection';
-import WikiaMapComponent from './wikia-map';
+import WikiaMapComponent from './article-media-map-thumbnail';
 import PortableInfoboxComponent from './portable-infobox';
 import AdsMixin from '../mixins/ads';
 import PollDaddyMixin from '../mixins/poll-daddy';
@@ -80,7 +80,6 @@ export default Ember.Component.extend(
 					this.loadIcons();
 					this.loadTableOfContentsData();
 					this.handleTables();
-					this.replaceMapsWithMapComponents();
 					this.replaceMediaPlaceholdersWithMediaComponents(this.get('media'), 4);
 					this.replaceImageCollectionPlaceholdersWithComponents(this.get('media'));
 					this.replaceWikiaWidgetsWithComponents();
@@ -147,7 +146,7 @@ export default Ember.Component.extend(
 			 * @returns {void}
 			 */
 			openLightbox(lightboxType, lightboxData) {
-				this.sendAction('openLightbox', lightboxType, lightboxData);
+				this.get('openLightbox')(lightboxType, lightboxData);
 			},
 
 			/**
@@ -245,6 +244,12 @@ export default Ember.Component.extend(
 						items: media[attrs.ref]
 					});
 				}
+			}
+
+			if (name === 'article-media-map-thumbnail') {
+				attrs = Ember.$.extend(attrs, {
+					openLightbox: this.get('openLightbox')
+				});
 			}
 
 			return {name, attrs, element};
@@ -412,41 +417,6 @@ export default Ember.Component.extend(
 			}
 
 			model.set('media', articleMedia);
-		},
-
-		/**
-		 * @returns {void}
-		 */
-		replaceMapsWithMapComponents() {
-			/**
-			 * @param {number} i
-			 * @param {Element} elem
-			 * @returns {void}
-			 */
-			this.$('.wikia-interactive-map-thumbnail').map((i, elem) => {
-				this.replaceMapWithMapComponent(elem);
-			});
-		},
-
-		/**
-		 * @param {Element} elem
-		 * @returns {void}
-		 */
-		replaceMapWithMapComponent(elem) {
-			const $mapPlaceholder = $(elem),
-				$a = $mapPlaceholder.children('a'),
-				$img = $a.children('img'),
-				mapComponent = this.createChildView(WikiaMapComponent.create({
-					url: $a.data('map-url'),
-					imageSrc: $img.data('src'),
-					id: $a.data('map-id'),
-					title: $a.data('map-title'),
-					click: 'openLightbox',
-				}));
-
-			mapComponent.createElement();
-			$mapPlaceholder.replaceWith(mapComponent.$());
-			mapComponent.trigger('didInsertElement');
 		},
 
 		/**
