+++ front/main/app/styles/component/_article-media.scss
@@ -50,6 +50,38 @@
 	}
 }
 
+.article-media-map-thumbnail {
+  margin: 40px 0;
+  position: relative;
+  text-align: center;
+
+  a {
+	display: block;
+  }
+
+  .icon {
+	fill: #fff;
+	position: absolute;
+	top: 5%;
+	right: 5%;
+  }
+
+  figcaption {
+	bottom: 0;
+	color: #fff;
+	font-size: 1.6rem;
+	font-weight: 700;
+	height: 50px;
+	left: 5%;
+	overflow: hidden;
+	padding: 5px 0;
+	position: absolute;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	width: 90%;
+  }
+}
+
 .article-media-gallery {
 	-webkit-overflow-scrolling: touch;
 	overflow: {
