--- front/main/app/styles/module/_wikia-map.scss
@@ -1,50 +1,0 @@
-.wikia-interactive-map-link {
-	pointer-events: none;
-}
-
-.wikia-map {
-	color: white;
-	height: auto;
-	position: relative;
-	width: 100%;
-
-	figcaption {
-		bottom: 0;
-		color: white;
-		font-size: 1.6rem;
-		font-weight: bold;
-		left: 5%;
-		overflow: hidden;
-		position: absolute;
-		text-overflow: ellipsis;
-		white-space: nowrap;
-		width: 90%;
-	}
-
-	.view-map-wrapper {
-		background: linear-gradient(to bottom, rgba(0,0,0,0) 0%,rgba(0,0,0,0.75) 100%);
-		height: 100%;
-		left: 0;
-		position: absolute;
-		top: 0;
-		width: 100%;
-	}
-
-	.view-map {
-		border: 1px solid white;
-		border-radius: 3px;
-		left: 50%;
-		line-height: 1rem;
-		padding: 8px 11px;
-		position: absolute;
-		top: 50%;
-		transform: translate(-50%, -50%);
-
-		svg {
-			fill: white;
-			width: 16px;
-			height: 16px;
-			vertical-align: bottom;
-		}
-	}
-}
