+++ front/main/app/styles/module/index.scss
@@ -19,6 +19,5 @@
 @import 'edit/edit';
 @import 'loader';
 @import 'site-footer';
-@import 'wikia-map';
 @import 'wikia-widget';
 @import 'wiki/index';
