+++ front/main/app/templates/components/article-media-map-thumbnail.hbs
@@ -1,0 +1,7 @@
+<a href="{{url}}">
+	<img src="{{imageSrc}}">
+	{{svg 'view-map' viewBox='0 0 20 20' class='icon'}}
+</a>
+{{#if caption}}
+    <figcaption>{{{caption}}}</figcaption>
+{{/if}}
