+++ front/main/app/templates/components/article-wrapper.hbs
@@ -51,7 +51,7 @@
 		content=model.content
 		media=model.media
 		noAds=noAds
-		openLightbox='openLightbox'
+		openLightbox=(action 'openLightbox')
 		updateHeaders='updateHeaders'
 		contributionEnabled=contributionEnabled
 		uploadFeatureEnabled=uploadFeatureEnabled
